fix heures stagiaires bis

+++ session/subscribers.php
@@ -93,11 +93,41 @@ if ($action == 'edit' && ($user->rights->agefodd->creer | $user->rights->agefodd
 		if ($agfsta->update($user) > 0) {
 
 			if (! empty($conf->global->AGF_USE_REAL_HOURS) && $agfsta->status_in_session !== 4) {
-
+			    
 				$agfssh = new Agefoddsessionstagiaireheures($db);
-				$agfssh->fetch_all_by_session(GETPOST('sessid', 'int'), GETPOST('stagerowid', 'int'));
+				$agfssh->fetch_all_by_session(GETPOST('sessid', 'int'), GETPOST('modstagid', 'int'));
 				foreach ( $agfssh->lines as $heures ) {
-					$heures->delete($user);
+				    if ($agfsta->status_in_session == 3 )
+				    {
+				        
+				        $TCal = array();
+				        if (!in_array($heures->fk_calendrier, $TCal))
+				        {
+				            $cal = new Agefodd_sesscalendar($db);
+				            $cal->fetch($heures->fk_calendrier);
+				            
+				            $TCal[$heures->fk_calendrier] = ($cal->heuref - $cal->heured) / 3600;
+				        }
+				        
+				        if (floatval($heures->heures) !== $TCal[$heures->fk_calendrier])
+				        {
+    				        $new_heures = new Agefoddsessionstagiaireheures($db);
+    				        
+    				        $new_heures->fk_stagiaire = $heures->fk_stagiaire;
+    				        $new_heures->nom_stagiaire = $heures->nom_stagiaire;
+    				        $new_heures->fk_user_author = $heures->fk_user_author;
+    				        $new_heures->fk_calendrier = $heures->fk_calendrier;
+    				        $new_heures->fk_session = $heures->fk_session;
+    				        $new_heures->heures = $TCal[$heures->fk_calendrier];
+    				        $new_heures->datec = $heures->datec;
+    				        
+    				        $res = $new_heures->create($user);
+    				        
+    				        if ($res) $heures->delete($user);
+    				        else setEventMessage($langs->trans('ErrUpdateHeures'), 'errors');
+				        }
+				    }
+					else $heures->delete($user);
 				}
 			}
